feat(core): override deselectAll and sortBy properly
[COM-1202]
