+++ packages/headless/src/controllers/facets/range-facet/numeric-facet/headless-numeric-facet.test.ts
@@ -23,6 +23,8 @@ import * as FacetIdDeterminor from '../../../core/facets/_common/facet-id-determ
 import {configuration, numericFacetSet, search} from '../../../../app/reducers';
 import {updateFacetOptions} from '../../../../features/facet-options/facet-options-actions';
 import {NumericFacetValue} from '../../../../features/facets/range-facets/numeric-facet-set/interfaces/response';
+import {deselectAllFacetValues} from '../../../../features/facets/facet-set/facet-set-actions';
+import {updateRangeFacetSortCriterion} from '../../../../features/facets/range-facets/generic/range-facet-actions';
 
 describe('numeric facet', () => {
   const facetId = '1';
@@ -108,6 +110,55 @@ describe('numeric facet', () => {
     it('dispatches a search', () => {
       const value = buildMockNumericFacetValue();
       numericFacet.toggleSelect(value);
+
+      const action = engine.actions.find(
+        (a) => a.type === executeSearch.pending.type
+      );
+      expect(action).toBeTruthy();
+    });
+  });
+
+  describe('#deselectAll', () => {
+    beforeEach(() => numericFacet.deselectAll());
+
+    it('dispatches #deselectAllFacetValues with the facet id', () => {
+      expect(engine.actions).toContainEqual(deselectAllFacetValues(facetId));
+    });
+
+    it('dispatches a #updateFacetOptions action with #freezeFacetOrder true', () => {
+      expect(engine.actions).toContainEqual(
+        updateFacetOptions({freezeFacetOrder: true})
+      );
+    });
+
+    it('dispatches a search', () => {
+      const action = engine.actions.find(
+        (a) => a.type === executeSearch.pending.type
+      );
+
+      expect(engine.actions).toContainEqual(action);
+    });
+  });
+
+  describe('#sortBy', () => {
+    it('dispatches #updateRangeFacetSortCriterion', () => {
+      const criterion = 'descending';
+      numericFacet.sortBy(criterion);
+      const action = updateRangeFacetSortCriterion({facetId, criterion});
+
+      expect(engine.actions).toContainEqual(action);
+    });
+
+    it('dispatches a #updateFacetOptions action with #freezeFacetOrder true', () => {
+      numericFacet.sortBy('descending');
+
+      expect(engine.actions).toContainEqual(
+        updateFacetOptions({freezeFacetOrder: true})
+      );
+    });
+
+    it('dispatches a search', () => {
+      numericFacet.sortBy('descending');
 
       const action = engine.actions.find(
         (a) => a.type === executeSearch.pending.type
